+++ administration/src/cards/PdfFactory.ts
@@ -4,7 +4,7 @@ import { QrCode } from '../generated/card_pb'
 import { Region } from '../generated/graphql'
 import { PdfConfig } from '../project-configs/getProjectConfig'
 import getDeepLinkFromQrCode from '../util/getDeepLinkFromQrCode'
-import { CardBlueprint } from './CardBlueprint'
+import { CardBlueprint } from './Card'
 import { CreateCardsResult } from './createCards'
 import pdfFormElement from './pdf/PdfFormElement'
 import pdfLinkArea from './pdf/PdfLinkArea'
