+++ administration/src/cards/__tests__/CsvFactory.test.ts
@@ -2,7 +2,7 @@ import { DynamicActivationCode } from '../../generated/card_pb'
 import { Region } from '../../generated/graphql'
 import bayernConfig from '../../project-configs/bayern/config'
 import nuernbergConfig from '../../project-configs/nuernberg/config'
-import { generateCardInfo, initializeCardBlueprint } from '../CardBlueprint'
+import { generateCardInfo, initializeCardBlueprint } from '../Card'
 import { CsvError, generateCsv, getCSVFilename } from '../CsvFactory'
 import { CreateCardsResult } from '../createCards'
 import { NUERNBERG_PASS_ID_EXTENSION_NAME } from '../extensions/NuernbergPassIdExtension'
