+++ administration/src/cards/pdf/PdfTextElement.ts
@@ -2,7 +2,7 @@ import { Color, PDFFont, PDFPage, RotationTypes } from 'pdf-lib'
 
 import { CardInfo } from '../../generated/card_pb'
 import { Region } from '../../generated/graphql'
-import { CardBlueprint } from '../CardBlueprint'
+import { CardBlueprint } from '../Card'
 import { Coordinates, PdfElement, mmToPt } from './PdfElements'
 
 export type InfoParams = {
